requirements.txt
altgraph==0.17.3
async-generator==1.10
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==3.2.0
chromedriver-autoinstaller==0.6.2
cryptography==37.0.2
et-xmlfile==1.1.0
execnet==2.0.2
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
macholib==1.16.2
numpy==1.21.6
openpyxl==3.1.2
outcome==1.2.0
packaging==21.3
pandas==1.3.5
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
pyOpenSSL==22.0.0
pyotp==2.8.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-json-report==1.5.0
pytest-metadata==3.0.0
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
selenium==4.6.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.9
webdriver-manager==4.0.0
wsproto==1.1.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zipp==3.8.0
